API asynctask add

+++ app/src/main/java/com/example/tnote/DataClass/WeatherMainData.kt
@@ -1,7 +1,7 @@
 package com.example.tnote.DataClass
 
 data class WeatherMainData(
-    var temp : Float,
+    var temp : Double,
     var feels_like : Any,
     var temp_min : Any,
     var temp_max : Any
